Subshell: do redirections before forking.
To make sure that the redirections stay in effect during the whole
life of the subshell, including the EXIT traps, do the redirects
before forking, and undo them in the parent process immediately after.

A test is added to check the state of redirections during the EXIT
trap.

Additionally, the string representation of subshell commands is
changed so that redirections appear outside of the parenthesis.